"renamed help package to utility"
